Move settings to config namespace
