+++ app/config/packages/fideloper/proxy/config.php
@@ -1,5 +1,5 @@
 <?php
-return array(
+return [
 
     /*
     |--------------------------------------------------------------------------
@@ -26,7 +26,7 @@ return array(
     |
     */
 
-    'proxies' => array(
+    'proxies' => [
         '204.93.240.0',
         '204.93.177.0',
         '199.27.128.0',
@@ -46,6 +46,6 @@ return array(
         '2803:f800::',
         '2405:b500::',
         '2405:8100::',
-    )
+    ],
 
-);
+];
