+++ app/database/seeds/UsersTableSeeder.php
@@ -21,7 +21,7 @@ class UsersTableSeeder extends Seeder
                 "password" => "test123",
                 "email"    => "test@test.com",
                 "level"    => 1,
-            ]
+            ],
         ];
 
         User::truncate();
